some requested changes, styling mostly
