+++ solution/ui/regulations/utilities/api.js
@@ -334,7 +334,25 @@ const getLastUpdatedDates = async (apiUrl, titleArr = ["42"]) => {
         Object.entries(combinedResults).map((arr) => [arr[1].name, arr[1].date])
     );
 };
-
+/**
+ * Gets the three most recent resources of a type.
+ * @param {*} apiURL  -base url for the api
+ * @param {*} type  - type of resource, fr doc or not
+ * @returns 3 resources
+ */
+const getRecentResources = async (apiURL, { page = 1, pageSize = 3 }, type) => {
+    if(type!=="rules"){
+        return httpApiGetLegacy(
+            `${apiURL}resources/supplemental_content?page=${page}&page_size=${pageSize}&paginate=true`,
+            {}, // params, default
+            apiURL);
+    } 
+    return httpApiGetLegacy(
+        `${apiURL}resources/federal_register_docs?page=${page}&page_size=${pageSize}&paginate=true`,
+        {}, // params, default
+        apiURL
+    );
+};
 /**
  *
  * Fetches and formats list of parts to be used as dictionary
@@ -525,6 +543,7 @@ export {
     getFormattedPartsList,
     getLastUpdatedDates,
     getParts,
+    getRecentResources,
     getRegSearchResults,
     getSearchGovResources,
     getSectionsForPart,
